Changed global functions to actual functions and renamed
