+++ cogs/moderation.py
@@ -5,7 +5,7 @@ import sqlite3
 import os
 from discord.ext import commands
 from discord.utils import get
-from .GlobalFunctions import GlobalFunctions as GF
+from .lib import check_block, get_id, has_moderator
 
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
 db_path = os.path.join(BASE_DIR, "db.db")
@@ -56,14 +56,8 @@ class ModerationCog(commands.Cog):
 
         return milliseconds
 
-    async def has_moderator(ctx):
-        for role in ctx.author.roles:
-            if role.id == 779430696510160936  or ctx.author.id == 344666116456710144 or role.id == 769842662580027452:
-                return True
-        return False
-
     async def not_blocked(ctx):
-        return GF.check_block(ctx.author.id, ctx.command.name)
+        return check_block(ctx.author.id, ctx.command.name)
         #85,579
 
     @commands.command(name="purge")
@@ -162,7 +156,7 @@ class ModerationCog(commands.Cog):
     @commands.check(has_moderator)
     @commands.check(not_blocked)
     async def mute(self, ctx, member: discord.Member, length, *, reason):
-        role_id = await GF.get_id(ctx.guild, "Muted")
+        role_id = await get_id(ctx.guild, "Muted")
         role = get(member.guild.roles, id=role_id)
 
         await member.add_roles(role, reason=reason, atomic=True)
@@ -181,7 +175,7 @@ class ModerationCog(commands.Cog):
     @commands.check(has_moderator)
     @commands.check(not_blocked)
     async def unmute(self, ctx, member: discord.Member):
-        role_id = await GF.get_id(ctx.guild, "Muted")
+        role_id = await get_id(ctx.guild, "Muted")
         role = get(member.guild.roles, id=role_id)
 
         await member.remove_roles(role, reason = "time's up ")
@@ -213,7 +207,7 @@ class ModerationCog(commands.Cog):
     @commands.check(has_moderator)
     @commands.check(not_blocked)
     async def dungeon(self, ctx, member: discord.Member, *, reason):
-        role_id = await GF.get_id(ctx.guild, "Dungoned")
+        role_id = await get_id(ctx.guild, "Dungoned")
         role = get(member.guild.roles, id=role_id)
 
         await member.add_roles(role, reason=reason, atomic=True)
@@ -228,7 +222,7 @@ class ModerationCog(commands.Cog):
     @commands.check(has_moderator)
     @commands.check(not_blocked)
     async def release(self, ctx, member: discord.Member):
-        role_id = await GF.get_id(ctx.guild, "Dungoned")
+        role_id = await get_id(ctx.guild, "Dungoned")
         role = get(member.guild.roles, id=role_id)
 
         await member.remove_roles(role, reason = "released")
