+++ cogs/owner.py
@@ -4,7 +4,6 @@ import asyncio
 import sqlite3
 import os
 from .classes.UserAccount import UserAccount
-from .GlobalFunctions import GlobalFunctions as GF
 from discord.utils import get
 from discord import FFmpegPCMAudio
 from discord.ext import commands
